updated description of anonymous function
Hi, I updated the description to properly define the examples. "Traditional function" to me usually has a name, so when the documentation used "traditional function" I was a bit confused as it really wasn't and I thought it might also be confusing to most beginners who haven't encountered anonymous functions.